adjust default value for `--rocksdb.max-subcompactions`
Adjust default value for startup option `--rocksdb.max-subcompactions` from 1
to 2. This allows compactions jobs to be broken up into disjoint ranges which
can be processed in parallel.

+++ arangod/RocksDBEngine/RocksDBOptionFeature.h
@@ -76,7 +76,7 @@ class RocksDBOptionFeature final : public application_features::ApplicationFeatu
   uint64_t _maxBytesForLevelBase;
   double _maxBytesForLevelMultiplier;
   int32_t _maxBackgroundJobs;
-  uint64_t _maxSubcompactions;
+  uint32_t _maxSubcompactions;
   uint32_t _numThreadsHigh;
   uint32_t _numThreadsLow;
   uint64_t _targetFileSizeBase;
